refactor(abw): split writer brief modules
