+++ apps/ai-blog-writer/apps/backend/tests/test_editor_assist_architecture.py
@@ -147,6 +147,38 @@ def test_research_profile_remains_a_thin_compatible_facade():
     }.issubset(imported_modules)
 
 
+def test_writer_brief_remains_a_thin_compatible_facade():
+    feature_path = Path(__file__).parents[1] / "app" / "features" / "editor_assist"
+    facade_path = feature_path / "writer_brief.py"
+    source = facade_path.read_text()
+    module = ast.parse(source)
+
+    top_level_defs = [
+        node
+        for node in module.body
+        if isinstance(node, (ast.ClassDef, ast.FunctionDef, ast.AsyncFunctionDef))
+    ]
+    imported_modules = {
+        node.module
+        for node in module.body
+        if isinstance(node, ast.ImportFrom) and node.module
+    }
+
+    assert len(source.splitlines()) <= 120
+    assert [node.name for node in top_level_defs] == [
+        "_invoke_curator_model",
+        "run_writer_brief",
+    ]
+    assert {
+        "writer_brief_contracts",
+        "writer_brief_execution",
+        "writer_brief_parsing",
+        "writer_brief_policy",
+        "writer_brief_prompt",
+        "writer_brief_rendering",
+    }.issubset(imported_modules)
+
+
 def test_editor_assist_internals_do_not_depend_on_listicle_writer_facade():
     feature_path = Path(__file__).parents[1] / "app" / "features" / "editor_assist"
     facade_importers: list[str] = []
